Create EigenproblemEquationSystem and move eigen methods into it
