+++ framework/include/mfem/equation_systems/EquationSystem.h
@@ -22,8 +22,6 @@
 #include "ScaleIntegrator.h"
 #include "NLScaleIntegrator.h"
 
-class MFEMEigensolverBase;
-
 namespace Moose::MFEM
 {
 
@@ -53,8 +51,6 @@ public:
    * eliminated-variable constraints, and populate the true-DoF vectors used by the solve.
    */
   void FormSystem(mfem::BlockVector & trueX, mfem::BlockVector & trueRHS);
-  /// Build eigenproblem system, with essential boundary conditions accounted for
-  void BuildEigenproblemJacobian(mfem::BlockVector & trueX, mfem::OperatorHandle & massRHS);
   /// Compute residual y = Mu
   void Mult(const mfem::Vector & u, mfem::Vector & residual) const override;
   /// Get Jacobian at the provided vector of true DoFs of trial variables
@@ -62,10 +58,6 @@ public:
 
   /// Update variable from solution vector after solve
   virtual void SetTrialVariablesFromTrueVectors(const mfem::BlockVector & trueX) const;
-
-  /// Update eigenvectors from solution after eigensolve
-  void RecoverEigenproblemSolution(Moose::MFEM::GridFunctions & gridfunctions,
-                                   MFEMEigensolverBase * eigensolver);
 
   // Test variables are associated with linear forms,
   // whereas trial variables are associated with gridfunctions.
@@ -102,8 +94,6 @@ protected:
                                 mfem::Array<int> & global_ess_markers);
   /// Update all essentially constrained true DoF markers and values on boundaries
   virtual void ApplyEssentialBCs();
-  /// Mark external boundaries as essential for eigenproblem BC elimination
-  virtual void ApplyEigenEssentialBCs();
   /// Perform trivial eliminations of coupled variables lacking corresponding test variables
   virtual void EliminateCoupledVariables();
   /// Build linear forms and eliminate constrained DoFs
@@ -134,12 +124,6 @@ protected:
                                 mfem::BlockVector & trueRHS);
   /// Compute Jacobian matrix at the provided vector of true DoFs of trial variables
   void FormJacobianMatrix(const mfem::Vector & u);
-
-  /// Form HypreParMatrix matrix operator for the eigensolver with Dirichlet BC elimination.
-  virtual void FormEigenproblemMatrix(mfem::OperatorHandle & op);
-
-  /// Form mass matrix for the eigensolver with Dirichlet BC elimination.
-  void FormMassMatrix(mfem::OperatorHandle & op);
 
   /**
    * Template method for applying BilinearFormIntegrators on domains from kernels to a BilinearForm,
@@ -242,7 +226,6 @@ protected:
 private:
   friend class EquationSystemProblemOperator;
   friend class ::MFEMProblemSolve;
-  friend class EquationSystemEigenproblemOperator;
   /// Disallowed inherited method
   using mfem::Operator::RecoverFEMSolution;
 };
