+++ framework/include/mfem/problem_operators/EquationSystemEigenproblemOperator.h
@@ -12,6 +12,7 @@
 #pragma once
 
 #include "EquationSystemProblemOperator.h"
+#include "EigenproblemEquationSystem.h"
 
 namespace Moose::MFEM
 {
@@ -19,11 +20,20 @@ namespace Moose::MFEM
 class EquationSystemEigenproblemOperator : public EquationSystemProblemOperator
 {
 public:
-  EquationSystemEigenproblemOperator(MFEMProblem & problem) : EquationSystemProblemOperator(problem)
+  EquationSystemEigenproblemOperator(MFEMProblem & problem)
+    : EquationSystemProblemOperator(problem)
   {
+    mooseAssert(std::dynamic_pointer_cast<EigenproblemEquationSystem>(_problem_data.eqn_system),
+                "MFEMProblemData::eqn_system is not an EigenproblemEquationSystem.");
   }
 
   virtual void Solve() override;
+
+  [[nodiscard]] EigenproblemEquationSystem * GetEquationSystem() const override
+  {
+    return static_cast<EigenproblemEquationSystem *>(
+        EquationSystemProblemOperator::GetEquationSystem());
+  }
 
   mfem::OperatorHandle _mass_rhs;
