+++ framework/src/mfem/equation_systems/EquationSystem.C
@@ -10,7 +10,6 @@
 #ifdef MOOSE_MFEM_ENABLED
 
 #include "EquationSystem.h"
-#include "MFEMEigensolverBase.h"
 #include "libmesh/int_range.h"
 
 namespace Moose::MFEM
@@ -249,20 +248,6 @@ EquationSystem::ApplyEssentialBCs()
 }
 
 void
-EquationSystem::ApplyEigenEssentialBCs()
-{
-  for (const auto i : index_range(_trial_var_names))
-  {
-    mfem::ParGridFunction & trial_gf = *(_var_ess_constraints.at(i));
-    _global_ess_markers.SetSize(trial_gf.ParFESpace()->GetParMesh()->bdr_attributes.Max());
-    _global_ess_markers = 0;
-    trial_gf.Update();
-    trial_gf = _gfuncs->GetRef(_trial_var_names.at(i));
-    trial_gf.ParFESpace()->GetParMesh()->MarkExternalBoundaries(_global_ess_markers);
-  }
-}
-
-void
 EquationSystem::EliminateCoupledVariables()
 {
   for (const auto & test_var_name : _test_var_names)
@@ -291,36 +276,6 @@ EquationSystem::FormLinearSystem(mfem::OperatorHandle & op,
                 "Non-legacy assembly is only supported for single test and trial variable systems");
     FormSystemOperator(op, trueX, trueRHS);
   }
-}
-
-void
-EquationSystem::FormEigenproblemMatrix(mfem::OperatorHandle & op)
-{
-  auto & test_var_name = _test_var_names.at(0);
-  auto blf = _blfs.Get(test_var_name);
-
-  blf->EliminateEssentialBCDiag(_global_ess_markers, 1.0);
-  blf->Finalize();
-  op.Reset(blf->ParallelAssemble());
-}
-
-void
-EquationSystem::FormMassMatrix(mfem::OperatorHandle & op)
-{
-  mfem::ConstantCoefficient one(1.0);
-  mfem::ParFiniteElementSpace * fespace = _test_pfespaces.at(0);
-  mfem::ParBilinearForm * m = new mfem::ParBilinearForm(fespace);
-
-  if (fespace->GetTypicalFE()->GetRangeType() == mfem::FiniteElement::SCALAR)
-    m->AddDomainIntegrator(new mfem::MassIntegrator(one));
-  else
-    m->AddDomainIntegrator(new mfem::VectorFEMassIntegrator(one));
-
-  m->Assemble();
-  // Shift the eigenvalue corresponding to eliminated dofs to a large value
-  m->EliminateEssentialBCDiag(_global_ess_markers, std::numeric_limits<mfem::real_t>::min());
-  m->Finalize();
-  op.Reset(m->ParallelAssemble());
 }
 
 void
@@ -429,20 +384,6 @@ EquationSystem::FormSystem(mfem::BlockVector & trueX, mfem::BlockVector & trueRH
 }
 
 void
-EquationSystem::BuildEigenproblemJacobian(mfem::BlockVector & trueX, mfem::OperatorHandle & massRHS)
-{
-  mooseAssert(_test_var_names.size() == 1 && (_test_var_names.size() == _trial_var_names.size()) &&
-                  (_test_var_names.at(0) == _trial_var_names.at(0)),
-              "Eigensolve is only supported for single-variable, square systems");
-
-  height = trueX.Size();
-  width = trueX.Size();
-  ApplyEigenEssentialBCs();
-  FormEigenproblemMatrix(_jacobian);
-  FormMassMatrix(massRHS);
-}
-
-void
 EquationSystem::Mult(const mfem::Vector & sol, mfem::Vector & residual) const
 {
   // Update gridfunctions that may be referenced by coefficients within nonlinear integrators
@@ -521,21 +462,6 @@ EquationSystem::SetTrialVariablesFromTrueVectors(const mfem::BlockVector & trueX
     auto & trial_var_name = _trial_var_names.at(i);
     trueX.GetBlock(i).SyncAliasMemory(trueX);
     _gfuncs->Get(trial_var_name)->Distribute(&(trueX.GetBlock(i)));
-  }
-}
-
-void
-EquationSystem::RecoverEigenproblemSolution(Moose::MFEM::GridFunctions & gridfunctions,
-                                            MFEMEigensolverBase * eigensolver)
-{
-  mfem::Array<mfem::real_t> eigenvalues;
-  eigensolver->getEigenvalues(eigenvalues);
-
-  for (int i = 0; i < eigenvalues.Size(); ++i)
-  {
-    auto & trial_var_name = _trial_var_names.at(0);
-    gridfunctions.Get(trial_var_name + "_" + std::to_string(i))
-        ->Distribute(eigensolver->getEigenvector(i));
   }
 }
